# Speedup some of the test setup