--- a/molecule/default/Dockerfile.j2
+++ b/molecule/default/Dockerfile.j2
@@ -56,5 +56,9 @@ RUN useradd --create-home --groups sudo --shell /bin/bash podman && \
     mkdir -p /var/lib/systemd/linger && \
     touch /var/lib/systemd/linger/podman
 
+# Configure podman
+COPY image/isolated_hosts /etc/isolated_hosts
+COPY image/containers.conf /etc/containers/containers.conf
+
 VOLUME /home/podman/.local/share/containers
 VOLUME /home/podman/infrastructure