--- a/molecule/default/image/containers.conf
+++ b/molecule/default/image/containers.conf
@@ -1,0 +1,2 @@
+[containers]
+base_hosts_file = "/etc/isolated_hosts"