--- a/molecule/default/image/isolated_hosts
+++ b/molecule/default/image/isolated_hosts
@@ -1,0 +1,2 @@
+127.0.0.1   localhost localhost.localdomain localhost4 localhost4.localdomain4
+::1         localhost localhost.localdomain localhost6 localhost6.localdomain6