--- a/pyproject.toml
+++ b/pyproject.toml
@@ -108,7 +108,7 @@ pylint.max-branches = 13
 # Tests
 ##
 [tool.pytest.ini_options]
-addopts = "--color=yes --verbose --verbose"
+addopts = "--color=yes --verbose --verbose --dist=loadgroup --numprocesses 4"
 
 filterwarnings = [
     "error",