--- a/requirements/requirements-tests.txt
+++ b/requirements/requirements-tests.txt
@@ -1,5 +1,6 @@
 Pygments
 pytest
+pytest-xdist
 pytest-testinfra
 pyyaml
 requests