- Parallelize some testing operations, reducing the time it takes to run the pytest checks by half (20s)
- Stop cleaning up the space before we run, there's apparently enough space now (15s)
- Install less dependencies in the base test image (5s)
- Don't use ansible to apply the podman config in the test image (5s)